Remove unsafeFlags
Unsafe flags break numbered releases.

Closes #32 again:
The issue I had was that I've added TimelaneCombine to both the app
target as well as the extension target. This led to TimelaneCombine
being built for the whole app instead of just with the extension api.
When including TimelaneCombine only into the extension everything works
flawlessly.
Both the app and the extension can be profiled.